Trait kvs::KvsEngine

pub trait KvsEngine {
    fn set(&mut self, key: String, value: String) -> Result<()>;

    fn get(&mut self, key: String) -> Result<Option<String>>;

    fn remove(&mut self, key: String) -> Result<()>;
}

Trait for a key value storage engine.

Required methods

fn set(&mut self, key: String, value: String) -> Result<()>

Sets the value of a string key to a string.

If the key already exists, the previous value will be overwritten.

fn get(&mut self, key: String) -> Result<Option<String>>

Gets the string value of a given string key.

Returns None if the given key does not exist.

fn remove(&mut self, key: String) -> Result<()>

Removes a given key.

Errors

It returns KvsError::KeyNotFound if the given key is not found.

Loading content...